rename preavis_externes to differenciate it from affaire_component
